chore(slack): rename eventDedup -> event-dedup to match kebab-case convention
Aligns the Slack event idempotency module filename with the rest of
src/slack/ which uses kebab-case (channel-activity.ts, thread-buffer.ts).

Addresses review feedback on PR #396.
